requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
decorator==5.1.1
distlib==0.3.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-autoslug==1.9.8
django-extensions==3.1.5
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
factory-boy==3.2.1
Faker==13.3.4
filelock==3.6.0
flake8==4.0.1
gunicorn==20.1.0
identify==2.4.12
idna==3.3
inflection==0.5.1
ipython==8.2.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
nose==1.3.7
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit==2.18.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
static3==0.7.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.1
whitenoise==6.0.0
